Package library

Package library

Requests to SAOImage/DS9

DS9.get — Function.

DS9.get([T, [dims,]] args...)

sends a "get" request to the SAOImage/DS9 server. The request is made of arguments args... converted into strings and merged with separating spaces. An exception is thrown in case of error.

The returned value depends on the optional arguments T and dims:

If neither T nor dims are specified, an instance of XPA.Reply is returned with at most one answer (see XPA.get for more details).

If only T is specified, it can be:

String to return the answer as a single string;

Vector{String}} or Tuple{Vararg{String}} to return the answer split in words as a vector or as a tuple of strings;

T where T<:Real to return a value of type T obtained by parsing the textual answer.

Tuple{Vararg{T}} where T<:Real to return a value of type T obtained by parsing the textual answer;

Vector{T} where T is not String to return the binary contents of the answer as a vector of type T;

If both T and dims are specified, T can be an array type like Array{S} or Array{S,N} and dims a list of N dimensions to retrieve the binary contents of the answer as an array of type Array{S,N}.

As a special case:

DS9.get(Array; endian=:native) -> arr

yields the contents of current SAOImage/DS9 frame as an array (or as nothing if the frame is empty). Keyword endian can be used to specify the byte order of the received values (see DS9.byte_order.

To retrieve the version of the SAOImage/DS9 program:

DS9.get(VersionNumber)

See also DS9.connect, DS9.set and XPA.get.

source

DS9.set — Function.

DS9.set(args...; data=nothing)

sends command and/or data to the SAOImage/DS9 server. The command is made of arguments args... converted into strings and merged with a separating spaces. Keyword data can be used to specify the data to send. An exception is thrown in case of error.

As a special case:

DS9.set(arr; mask=false, new=false, endian=:native)

set the contents of the current SAOImage/DS9 frame to be array arr. Keyword new can be set true to create a new frame for displyaing the array. Keyword endian can be used to specify the byte order of the values in arr (see DS9.byte_order.

See also DS9.connect, DS9.get and XPA.set.

source

Connection

DS9.connect — Function.

DS9.connect(apt="DS9:*") -> ident

set the access point for further SAOImage/DS9 commands. Argument apt identifies the XPA access point, it can be a template string like "DS9:*" which is the default value. The returned value is the name of the access point.

To retrieve the name of the current SAOImage/DS9 access point, call the DS9.connection method.

See also DS9.accesspoint and DS9.connection.

source

DS9.accesspoint — Function.

DS9.accesspoint()

yields the XPA access point which identifies the SAOImage/DS9 server. This access point can be set by calling the DS9.connection method. An empty string is returned if no access point has been chosen. To automatically connect to SAOImage/DS9 if not yet done, you can do:

if DS9.accesspoint() == ""; DS9.connect(); end

See also DS9.connect and DS9.connection.

source

DS9.connection — Function.

DS9.connection()

yields the XPA persistent client connection used to communicate with SAOImage/DS9 server(s).

See also DS9.accesspoint, DS9.connect and XPA.Client.

source

Utilities

DS9.bitpix_of — Function.

DS9.bitpix_of(x) -> bp

yields FITS bits-per-pixel (BITPIX) value for x which can be an array or a type. A value of 0 is returned if x is not of a supported type.

See also DS9.bitpix_to_type.

source

DS9.bitpix_to_type — Function.

DS9.bitpix_to_type(bp) -> T

yields Julia type corresponding to FITS bits-per-pixel (BITPIX) value bp. The value Nothing is returned if bp is unknown.

See also DS9.bitpix_of.

source

DS9.byte_order — Function.

DS9.byte_order(endian)

yields the byte order for retrieving the elements of a SAOImage/DS9 array. Argument can be one of the strings (or the equivalent symbol): "big" for most significant byte first, "little" for least significant byte first or "native" to yield the byte order of the machine.

See also DS9.get, DS9.set.

source